- Delete Float extension - Add Double extension - Create account and transaction table view cell classes and their view models

+++ Bank Account App/Bank Account App/Extensions/DoubleExtension.swift
@@ -8,7 +8,7 @@
 
 import Foundation
 
-extension Float
+extension Double
 {
     var dollarsFormatting:String
     {
